[CI] Refactor CI
