+++ tests/long_term/spec_decode/e2e/test_v1_spec_decode.py
@@ -1,14 +1,11 @@
 # SPDX-License-Identifier: Apache-2.0
 from __future__ import annotations
 
-import os
 import random
 from typing import Any
 
 import pytest
 from vllm import LLM, SamplingParams
-
-os.environ["VLLM_USE_MODELSCOPE"] = "True"
 
 
 @pytest.fixture
