+++ tests/model_utils.py
@@ -20,9 +20,6 @@
 import warnings
 from typing import Callable, Dict, List, Optional, Sequence, Tuple, Union
 
-import torch
-from vllm.config import ModelConfig, TaskOption
-from vllm.inputs import InputContext
 from vllm.sequence import Logprob, PromptLogprobs, SampleLogprobs
 
 TokensText = Tuple[List[int], str]
@@ -262,45 +259,6 @@ def check_logprobs_close(
                     warnings.simplefilter("always")
 
                     warnings.warn(fail_msg, stacklevel=2)
-
-
-def build_model_context(model_name: str,
-                        task: TaskOption = "auto",
-                        tokenizer_name: Optional[str] = None,
-                        trust_remote_code: bool = False,
-                        dtype: Optional[Union[str, torch.dtype]] = None,
-                        mm_processor_kwargs: Optional[Dict] = None,
-                        limit_mm_per_prompt: Optional[Dict] = None):
-    """Creates an InputContext for a given model.
-
-    Args:
-        model_name: Name of the model being considered.
-        tokenizer_name: Name of the tokenizer being considered.
-        trust_remote_code: Whether or not to allow loading remote code.
-        mm_processor_kwargs: optional processor kwargs for to be leveraged
-            in the input processor, mapper, dummy data creation, etc.
-        limit_mm_per_prompt: Multimodal limits.
-
-    Returns:
-        InputContext for the model being considered.
-    """
-    if tokenizer_name is None:
-        tokenizer_name = model_name
-    if dtype is None:
-        dtype = "half"
-
-    model_config = ModelConfig(
-        model_name,
-        task=task,
-        tokenizer=tokenizer_name,
-        tokenizer_mode="auto",
-        trust_remote_code=trust_remote_code,
-        dtype=dtype,
-        seed=0,
-        mm_processor_kwargs=mm_processor_kwargs,
-        limit_mm_per_prompt=limit_mm_per_prompt,
-    )
-    return InputContext(model_config)
 
 
 def qwen_prompt(questions: List[str]) -> List[str]:
