+++ tests/multicard/test_offline_inference_distributed.py
@@ -28,15 +28,9 @@ import vllm  # noqa: F401
 from tests.conftest import VllmRunner
 
 os.environ["PYTORCH_NPU_ALLOC_CONF"] = "max_split_size_mb:256"
-os.environ["VLLM_USE_MODELSCOPE"] = "True"
 
 
-@pytest.mark.parametrize("model, distributed_executor_backend", [
-    ("Qwen/QwQ-32B", "mp"),
-    ("deepseek-ai/DeepSeek-V2-Lite", "mp"),
-])
-def test_models_distributed(model: str,
-                            distributed_executor_backend: str) -> None:
+def test_models_distributed_QwQ():
     example_prompts = [
         "vLLM is a high-throughput and memory-efficient inference and serving engine for LLMs.",
         "Briefly describe the major milestones in the development of artificial intelligence from 1950 to 2020.",
@@ -45,14 +39,28 @@ def test_models_distributed(model: str,
     dtype = "half"
     max_tokens = 5
     with VllmRunner(
-            model,
+            "Qwen/QwQ-32B",
             dtype=dtype,
             tensor_parallel_size=4,
-            distributed_executor_backend=distributed_executor_backend,
+            distributed_executor_backend="mp",
     ) as vllm_model:
         vllm_model.generate_greedy(example_prompts, max_tokens)
 
 
-if __name__ == "__main__":
-    import pytest
-    pytest.main([__file__])
+@pytest.mark.skipif(os.getenv("VLLM_USE_V1") == "1",
+                    reason="deepseek v2 lite is not supported on v1")
+def test_models_distributed_DeepSeek():
+    example_prompts = [
+        "vLLM is a high-throughput and memory-efficient inference and serving engine for LLMs.",
+        "Briefly describe the major milestones in the development of artificial intelligence from 1950 to 2020.",
+        "Compare and contrast artificial intelligence with human intelligence in terms of processing information.",
+    ]
+    dtype = "half"
+    max_tokens = 5
+    with VllmRunner(
+            "deepseek-ai/DeepSeek-V2-Lite",
+            dtype=dtype,
+            tensor_parallel_size=4,
+            distributed_executor_backend="mp",
+    ) as vllm_model:
+        vllm_model.generate_greedy(example_prompts, max_tokens)
