+++ tests/singlecard/ops/test_rotary_embedding.py
